Initialize TR and num meas values from an existing design matrix.

+++ src/qt/DesignEditor.h
@@ -18,7 +18,7 @@ class DesignEditor : public QWizard {
 
   void handleConditionClick(QMouseEvent*);
 
-  void setRepTime(int);
+  void setRepTime(double);
 
   void setNumMeas(int);
